Projek Phyton bot Telegram Phyton/requirements.txt
python-telegram-bot
pytesseract
Pillow
flask